pyproject.toml
[project]
name = "beauty-store-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-ninja>=1.4.5",
]
